failure-screenshots/libguides/bobcat-redirects/: delete obsolete screenshots from the initial multiple run checkin [+]
The initial batch was a merging of multiple runs.  There were two test cases which failed in earlier runs that succeeded in later ones, and when the merging of screenshots were done there was no pruning of obsolete PNG files, so pruning now by checking against index.json, which should be correct because the merge was a purely additive process.